sdk_common -> common
