+++ common/__init__.py
@@ -3,3 +3,6 @@
 #
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
+from dotenv import load_dotenv
+
+load_dotenv()
