+++ examples/custom_tools/ticker_data.py
@@ -9,8 +9,8 @@ import sys
 from typing import Dict
 
 import yfinance as yf
+from common.custom_tools import SingleMessageCustomTool
 from llama_stack.types.tool_param_definition_param import ToolParamDefinitionParam
-from sdk_common.custom_tools import SingleMessageCustomTool
 
 
 class TickerDataTool(SingleMessageCustomTool):
