+++ examples/scripts/hello.py
@@ -10,19 +10,13 @@
 import asyncio
 
 import fire
+from common.agents.event_logger import EventLogger
+from common.client_utils import *  # noqa: F403
 
 from llama_stack import LlamaStack
 
 from llama_stack.types import SamplingParams, UserMessage
 from llama_stack.types.agent_create_params import AgentConfig
-from sdk_common.agents.event_logger import EventLogger
-
-from sdk_common.client_utils import (
-    load_api_keys_from_env,
-    make_agent_config_with_custom_tools,
-    QuickToolConfig,
-    search_tool_defn,
-)
 from termcolor import cprint
 
 from .multi_turn import execute_turns, prompt_to_turn
