+++ examples/scripts/inflation.py
@@ -14,22 +14,9 @@ from examples.custom_tools.ticker_data import TickerDataTool
 from llama_stack import LlamaStack
 
 from llama_stack.types import Attachment, SamplingParams, UserMessage
-from llama_stack.types.agent_create_params import (
-    AgentConfig,
-    AgentConfigToolCodeInterpreterToolDefinition,
-    AgentConfigToolFunctionCallToolDefinition,
-    AgentConfigToolMemoryToolDefinition,
-    AgentConfigToolSearchToolDefinition,
-    AgentConfigToolWolframAlphaToolDefinition,
-)
-from sdk_common.agents.event_logger import EventLogger
-from sdk_common.client_utils import (
-    AttachmentBehavior,
-    load_api_keys_from_env,
-    make_agent_config_with_custom_tools,
-    QuickToolConfig,
-    search_tool_defn,
-)
+from llama_stack.types.agent_create_params import *  # noqa: F403
+from common.agents.event_logger import EventLogger
+from common.client_utils import *  # noqa: F403
 from termcolor import cprint
 
 from .multi_turn import execute_turns, prompt_to_turn
