+++ examples/scripts/llama_guard_demo.py
@@ -11,12 +11,7 @@ import asyncio
 
 import fire
 
-from sdk_common.client_utils import (
-    load_api_keys_from_env,
-    make_agent_config_with_custom_tools,
-    QuickToolConfig,
-    search_tool_defn,
-)
+from common.client_utils import *  # noqa: F403
 
 from .multi_turn import execute_turns, prompt_to_turn
