+++ requirements.txt
@@ -1,12 +1,10 @@
 fastapi
 fire
 httpx
-llama-stack>=0.0.20
 mesop
 python-dotenv
 setuptools
 uvicorn
 yfinance
-llama-stack>=0.0.19
 gradio
 llama-stack-client
